consertado erro quando adicionado código após a última linha `end.`; usando C++23; outras mudanças

+++ include/Pascal--/lexical/Lexeme.hpp
@@ -1,17 +1,22 @@
 #ifndef LEXEME_HPP
 #define LEXEME_HPP
 
+#include <cstdint>
 #include <string>
 #include "Pascal--/lexical/TokenType.hpp"
 
 struct Lexeme {
     std::string token;
     enum TokenType type;
-    int line;
-    int column;
+    uint16_t line;
+    uint16_t column;
 
     Lexeme();
-    Lexeme(const std::string& token, enum TokenType type, int line, int column);
+    Lexeme(
+        const std::string& token,
+        enum TokenType type,
+        uint16_t line, uint16_t column
+    );
 
     virtual ~Lexeme();
